[OM-26] Changing queries to unique file
