+++ lib/api/firebase.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:http/http.dart';
 import 'package:omni_manager/api/auth.dart';
@@ -8,6 +9,23 @@ final CollectionReference _metricsCollection = _firestore.collection('Metrics');
 
 class Database {
   static String? userUid = getUserUid();
+
+  static Future<bool> validateManager(Map<String, String> managerData) async {
+  CollectionReference users = FirebaseFirestore.instance.collection("Users");
+  // flow: vai encontrar o doc com email, determinar se existe e retornar um bool
+  return users
+      .where("email", isEqualTo: managerData["email"])
+      .where("company", isEqualTo: managerData["company"])
+      .where("department", isEqualTo: managerData["department"])
+      .limit(1)
+      .get()
+      .then((snapshot) {
+    if (snapshot.docs.isNotEmpty) {
+      return true;
+    }
+    return false;
+  });
+}
 
   static Stream<QuerySnapshot> listRatings() {
     return _metricsCollection
@@ -62,5 +80,37 @@ class Database {
               })
             })
         .catchError((error) => print('error while fetching employees'));
+  }
+}
+
+// example
+class GetUserName extends StatelessWidget {
+  final String docID;
+
+  GetUserName(this.docID);
+
+  @override
+  Widget build(BuildContext context) {
+    CollectionReference users = FirebaseFirestore.instance.collection("Users");
+
+    return FutureBuilder<DocumentSnapshot>(
+        future: users.doc(this.docID).get(),
+        builder:
+            (BuildContext context, AsyncSnapshot<DocumentSnapshot> snapshot) {
+          if (snapshot.hasError) {
+            print("Snapshot error: ${snapshot.error}");
+            return Text("Something went wrong");
+          }
+          if (snapshot.hasData && !(snapshot.data!.exists)) {
+            return Text("No data has been retrieved");
+          }
+          if (snapshot.connectionState == ConnectionState.done) {
+            Map<String, dynamic> data =
+                snapshot.data!.data() as Map<String, dynamic>;
+            return Text("${data['name']}");
+          }
+
+          return Text("Loading...");
+        });
   }
 }
